requirements.txt
Django>=3.1,<3.2
wagtail>=2.12,<2.13
django-extensions==3.1.1
django-widget-tweaks==1.4.8
sentry-sdk-0.20.2